refactor: modularize map layers
